# Add write-ahead log