--- a/Makefile
+++ b/Makefile
@@ -1,0 +1,6 @@
+test:
+	go test -race ./...
+
+.PHONY: gen/proto
+gen/proto:
+	buf generate