--- a/buf.work.yaml
+++ b/buf.work.yaml
@@ -1,0 +1,3 @@
+version: v1
+directories:
+  - proto