… aggregate_test.go
… buf.gen.yaml
… db_test.go
… design/wal.md
… distinct_test.go
… examples/simple.go
… filter_test.go
… gen/proto/go/frostdb/schema/v1alpha1/schema.pb.go
… gen/proto/go/frostdb/schema/v1alpha1/schema_vtproto.pb.go
… gen/proto/go/frostdb/wal/v1alpha1/wal.pb.go
… gen/proto/go/frostdb/wal/v1alpha1/wal_vtproto.pb.go
… go.sum
… proto/buf.yaml
… wal/wal.go
… wal/wal_test.go