This adds a write-ahead log based on [tidwall/wal](). The whole design can be found in `design/wal.md`. As a side effect of having to replay schemas, we can now serialize and deserialize schemas using protobuf (and jsonpb).

The durability guarantees we get through this outweigh the minuscule performance overhead I think:

```
$ benchstat -delta-test=none ../old.txt ../new.txt
name                                          old time/op  new time/op  delta
_Table_Insert_1000Rows_10Iters_10Writers-10    603ms ± 0%   626ms ± 0%  +3.77%
_Table_Insert_100Rows_1000Iters_1Writers-10    569ms ± 0%   595ms ± 0%  +4.64%
_Table_Insert_100Rows_100Iters_100Writers-10   63.6s ± 0%   65.3s ± 0%  +2.75%
```

@metalmatze @thorfour