@brancz 2022-07-21
Ok so I made a mistake with the benchmarks, they didn't actually turn on the WAL.

Here are the benchmarks with the first commit of this PR (it couldn't even run all benchmarks):

```
$ benchstat -delta-test=none ../old.txt ../new.txt
name                                         old time/op  new time/op   delta
_Table_Insert_1000Rows_10Iters_10Writers-10   609ms ± 0%   2119ms ± 0%   +247.72%
_Table_Insert_100Rows_1000Iters_1Writers-10   540ms ± 0%  24732ms ± 0%  +4475.96%
```

The culprit is ultimately always the synchronization and actual writing of WAL records. Therefore I decided to relax the durability guarantee a bit and have records be written asynchronously in batches. Doing that makes the situation pretty much identical to what we had previously:

```
$ benchstat -delta-test=none ../old.txt ../new.txt
name                                          old time/op  new time/op  delta
_Table_Insert_1000Rows_10Iters_10Writers-10    609ms ± 0%   602ms ± 0%  -1.19%
_Table_Insert_100Rows_1000Iters_1Writers-10    540ms ± 0%   550ms ± 0%  +1.70%
_Table_Insert_100Rows_100Iters_100Writers-10   64.5s ± 0%   63.2s ± 0%  -2.09%
```