@thorfour 2022-07-21
> 

If it's not going to really be a WAL then these should be consistency settings that users can tweak such that they could have a fully guarnteed wal