@brancz 2022-07-21
> If it's not going to really be a WAL then these should be consistency settings that users can tweak such that they could have a fully guarnteed wal

Pretty much all WALs are best-effort, especially those that rely on POSIX fsync (which this one does as well), meaning the underlying file is flushed by the kernel, which can be up to 30s after the write was issued.

I would agree with allowing tweaking if the result of tweaking was in any shape or form reasonable, but exponential synchronization cost is never going to work, so I wouldn't want to allow that.